git-svn-id: https://plugins.svn.wordpress.org/facelift-image-replacement/trunk@259130 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ utility.class.php
@@ -101,7 +101,7 @@
                     $dir_handle = opendir($dirname);
                 } 
                 while($file = readdir($dir_handle)) { 
-                    if($file != "." && $file != ".." && $file != '.svn') { 
+                    if($file != "." && $file != ".." && $file != '.svn' && $file != 'index.html') { 
                         if(!is_dir($dirname."/".$file)){
                             unlink ($dirname."/".$file);
                         }
